requirements.txt
streamlit>=1.49.0
python-pptx>=1.0.2
Pillow>=11.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0